+++ app/page.tsx
@@ -1,6 +1,5 @@
 import HomeComponent from "./home/_components/HomeComponent/HomeComponent";
-import { redirect } from "next/navigation";
-import TokenHandler from "./(beforeLogin)/auth/_components/TokenHandler";
+import AuthRouter from "./(beforeLogin)/auth/_components/AuthRouter";
 
 type Props = {
   searchParams: Promise<{ oAuthId?: string; status?: string }>;
@@ -8,13 +7,8 @@ type Props = {
 
 export default async function Home({ searchParams }: Props) {
   const query = await searchParams;
-  if (query.status === "new") {
-    redirect("/auth/signup");
-  }
 
-  if (query.status === "registered" && query.oAuthId) {
-    return <TokenHandler oAuthId={query.oAuthId} />;
-  }
+  if (query) return <AuthRouter query={query} />;
 
   return <HomeComponent />;
 }
