+++ services/getTokens.ts
@@ -13,25 +13,3 @@ export async function getTokens(oAuthId: string) {
 
   return response.json();
 }
-
-export async function getRefreshToken() {
-  const accessToken = sessionStorage.getItem("accessToken");
-  const refreshToken = sessionStorage.getItem("refreshToken");
-
-  const response = await fetch(
-    `${process.env.NEXT_PUBLIC_BASE_URL}/v1/auth/tokens`,
-    {
-      method: "POST",
-      body: JSON.stringify({
-        accessToken,
-        refreshToken,
-      }),
-    },
-  );
-
-  if (!response.ok) {
-    throw new Error(response.statusText);
-  }
-
-  return response.json();
-}
